requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
requests
passlib[bcrypt]
pydantic
python-multipart
sqlalchemy